Delete warnings
- Changed warning level to /W4.
- Suppressed or fixed all build warnings.
- Added visual studio project.
    - Can be built using visual studio but may require changing nmake settings.
    - For a true rebuild, use clean first.

+++ src/colors.cpp
@@ -53,7 +53,7 @@ ItemProperties Colors::blend_heritage(std::multiset<Color> colors) {
     ItemProperties heritage_props;
 
     for (auto it = begin(colors); it != end(colors);) {
-        const ItemProperties &other_props = heritage_properties(*it, colors.count(*it));
+        const ItemProperties &other_props = heritage_properties(*it, (int)colors.count(*it));
         heritage_props.map.insert(std::begin(other_props), std::end(other_props));
         std::advance(it, colors.count(*it));
     }
